Make screen culler reusable
